preserve folder structure on upload
the download function was refactored to download dashboards using newer
models from grafana directly. these models are more compatible with the
newer upload api, which is not fully supported in the grafana-sdk.

adds a client library for interacting with folders
enables creation of folders via API using the .folder.json found in the
target dashboard directories.

the upload cmd was rehauled to preserve folder structure when uploading
dashboards. missing dashboards will be created, existing dashboards will
be updated. missing folders will be created, existing folders will be
updated or re-used.